backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
greenlet==3.0.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Async task processing
celery==5.3.4
redis==5.0.1
pika==1.3.2
celery-redbeat

# Trading and financial libraries
ccxt==4.1.77
pandas>=2.0.0
numpy>=1.24.0
scikit-learn==1.3.2
ta==0.10.2

# Data processing and analysis
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Date and time
python-dateutil==2.8.2
pytz==2023.3

# Logging and monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# WebSocket support
websockets==12.0

# Environment and configuration
python-dotenv==1.0.0

# Email support
fastapi-mail==1.4.1

# File handling
python-magic==0.4.27

# Rate limiting
slowapi==0.1.9

# CORS
fastapi-cors==0.0.6

# Added from the code block
requests>=2.31.0

# Added from the code block
pydantic-settings
python-binance
python-jose[cryptography]
redis
requests
SQLAlchemy
uvicorn[standard]
watchfiles
pandas
pandas-ta 